Move logic edit goal logic to goal header, change design to match Figma
